persist: Move to dedicated package
 * mongo: Specify a small connection timeout
 * Create dedicated package
 * Remove unnecessary logging condition
