+++ qtoggleserver/drivers/persist/json.py
@@ -6,7 +6,8 @@ import os
 from typing import Any, Dict, Iterable, List, Optional
 
 from qtoggleserver.conf import settings
-from qtoggleserver.persist import BaseDriver, Id, Record
+from qtoggleserver.persist import BaseDriver
+from qtoggleserver.persist.typing import Id, Record
 from qtoggleserver.utils import json as json_utils
