+++ qtoggleserver/persist/typing.py
@@ -1,0 +1,6 @@
+
+from typing import Any, Dict
+
+
+Id = str
+Record = Dict[str, Any]
